Improved install scripts for Ubuntu and Debian.

+++ ucrm_install_debian.sh
@@ -16,7 +16,7 @@ if [ $? = 1 ]; then
 fi
 
 echo "Download and install Docker compose"
-curl -L https://github.com/docker/compose/releases/download/1.7.0/docker-compose-`uname -s`-`uname -m` > /usr/local/bin/docker-compose
+curl -L https://github.com/docker/compose/releases/download/1.7.1/docker-compose-`uname -s`-`uname -m` > /usr/local/bin/docker-compose
 chmod +x /usr/local/bin/docker-compose
 
 if [ -z "$(getent passwd $UCRM_USER)" ]; then
@@ -25,18 +25,20 @@ if [ -z "$(getent passwd $UCRM_USER)" ]; then
 	usermod -aG docker $UCRM_USER
 fi
 
-echo "Downloading docker compose files"
-curl -o /home/$UCRM_USER/docker-compose.yml https://raw.githubusercontent.com/U-CRM/billing/master/docker-compose.yml
-curl -o /home/$UCRM_USER/docker-compose.env https://raw.githubusercontent.com/U-CRM/billing/master/docker-compose.env
+if [ ! -f /home/$UCRM_USER/docker-compose.yml ]; then
+	echo "Downloading docker compose files"
+	curl -o /home/$UCRM_USER/docker-compose.yml https://raw.githubusercontent.com/U-CRM/billing/master/docker-compose.yml
+	curl -o /home/$UCRM_USER/docker-compose.env https://raw.githubusercontent.com/U-CRM/billing/master/docker-compose.env
 
-echo "Replacing path in docker compose"
-sed -i -e "s/#volumes:/volumes:/g" /home/$UCRM_USER/docker-compose.yml
-sed -i -e "s/#  \- \/home\/docker\/ucrm\/postgres:\/var\/lib\/postgresql\/data/  - \/home\/$UCRM_USER\/postgres:\/var\/lib\/postgresql\/data/g" /home/$UCRM_USER/docker-compose.yml
-sed -i -e "s/#  \- \/home\/docker\/ucrm:\/data/  - \/home\/$UCRM_USER\/data:\/data/g" /home/$UCRM_USER/docker-compose.yml
+	echo "Replacing path in docker compose"
+	sed -i -e "s/#volumes:/volumes:/g" /home/$UCRM_USER/docker-compose.yml
+	sed -i -e "s/#  \- \/home\/docker\/ucrm\/postgres:\/var\/lib\/postgresql\/data/  - \/home\/$UCRM_USER\/postgres:\/var\/lib\/postgresql\/data/g" /home/$UCRM_USER/docker-compose.yml
+	sed -i -e "s/#  \- \/home\/docker\/ucrm:\/data/  - \/home\/$UCRM_USER\/data:\/data/g" /home/$UCRM_USER/docker-compose.yml
 
-echo "Replacing env in docker compose"
-sed -i -e "s/POSTGRES_PASSWORD=ucrmdbpass1/POSTGRES_PASSWORD=$POSTGRES_PASSWORD/g" /home/$UCRM_USER/docker-compose.env
-sed -i -e "s/SECRET=changeThisSecretKey/SECRET=$SECRET/g" /home/$UCRM_USER/docker-compose.env
+	echo "Replacing env in docker compose"
+	sed -i -e "s/POSTGRES_PASSWORD=ucrmdbpass1/POSTGRES_PASSWORD=$POSTGRES_PASSWORD/g" /home/$UCRM_USER/docker-compose.env
+	sed -i -e "s/SECRET=changeThisSecretKey/SECRET=$SECRET/g" /home/$UCRM_USER/docker-compose.env
+fi
 
 echo "Downloading docker images"
 cd /home/$UCRM_USER && /usr/local/bin/docker-compose pull
